backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
transformers==4.40.0
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
scikit-learn==1.4.2
pandas==2.2.2
sqlalchemy==2.0.29
python-multipart==0.0.9
openpyxl==3.1.2
requests==2.31.0
python-dateutil==2.9.0
regex==2024.4.16
apscheduler==3.10.4
numpy==1.26.4
joblib==1.4.0
